backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
psycopg[binary]>=3.1.0
bcrypt>=4.1.0
python-dotenv>=1.0.0
